modularized queries and routers based on suggestion from P
