+++ app/routes/users.py
@@ -1,0 +1,40 @@
+from queries import users
+import schemas
+from queries.users import db_get_user, db_get_users
+from fastapi import (
+    Depends,
+    HTTPException,
+    status,
+    Response,
+    APIRouter,
+    Request,
+    FastAPI,
+)
+from sqlalchemy.orm import Session
+from typing import List
+from database import SessionLocal
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+router = APIRouter()
+
+
+@router.get("/users", response_model=List[schemas.UserSchema])
+def get_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
+    users = db_get_users(db, skip=skip, limit=limit)
+    return users
+
+
+@router.get("/users/{user_id}", response_model=schemas.UserSchema)
+def get_user(user_id: int, db: Session = Depends(get_db)):
+    user = db_get_user(db, user_id)
+    if user is None:
+        raise HTTPException(status_code=400, detail="user not found")
+    return user
